Add workflow that runs GhostInspector test suite
This commit adds a reusable workflow that takes a GhostInspector test
suite ID, and a GhostInspector API key, then runs the test suite and
returns success if it has passed, and otherwise returns a failure.

There are a couple of things to note here:

- We are running the test suite afresh, rather than just checking the
  status of the last run. According to the GhostInspector API docs
  (https://ghostinspector.com/docs/api/suites/#execute) this can take up
  to 20 minutes in exceptional circumstances. Hence, the timeout on the
  CURL request is set to 1200 seconds.
- The API request to run the test suite, and the request to get the
  result are separate. So we run the test suite, wait for that to
  complete, then run another request to get the latest result for that
  suite.

There is a Docker image you can use to access Ghost Inspector CLI in
GitHub actions (see
https://ghostinspector.com/docs/integration/github-actions/), however, I
have opted not to use that here, because:

a) Passing args to Docker containers is not currently compatible with
act, which means the workflow would not be testable locally (see
https://github.com/nektos/act/issues/917)
b) Using a container to make a couple of CURL requests feels like
overkill, and would probably slow the workflow down. Making the requests
manually gives us more control over the output of the workflow.